requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
langchain==0.0.350
langchain-openai==0.0.2
langgraph==0.0.62
faiss-cpu==1.7.4
numpy==1.24.3
python-multipart==0.0.6